standardize testbench.cc location
