+++ MANIFEST.in
@@ -3,3 +3,4 @@ recursive-include switchboard/dpi *.h *.hh *.hpp *.c *.cc *.cpp
 recursive-include switchboard/vpi *.h *.hh *.hpp *.c *.cc *.cpp
 recursive-include switchboard/deps/verilog-axi/rtl *.v *.sv *.vh *.svh
 recursive-include switchboard/verilog *.v *.sv *.vh *.svh
+recursive-include switchboard/verilator *.h *.hh *.hpp *.c *.cc *.cpp
