+++ examples/router/test.py
@@ -39,13 +39,10 @@ def start_router(aq, bq, cq, dq):
 
 
 def build_testbench(fast=False):
-    dut = SbDut('testbench')
-
-    EX_DIR = Path('..').resolve()
+    dut = SbDut('testbench', default_main=True)
 
     # Set up inputs
     dut.input('testbench.sv')
-    dut.input(EX_DIR / 'common' / 'verilator' / 'testbench.cc')
 
     # Settings
     dut.set('option', 'trace', True)  # enable VCD (TODO: FST option)
